requirements.txt
pytest
boto3
Django==3.1.7
django-extensions==3.1.1
djangorestframework==3.12.2
django-storages==1.11.1
django-ses==1.0.3
django-allow-cidr==0.3.1
django-nose==1.4.7
pycountry==20.7.3
Pillow==8.1.0
django-filter==2.4.0
django-model-utils==4.1.1
stringcase==1.2.0
environs==9.3.1
dj-database-url==0.5.0
django-tenants==3.2.1
psycopg2==2.8.6
Pillow==8.1.0
colour-science==0.3.16
numpy==1.20.1
